requirements.txt
flask
flask-cors
tiktoken
openai
GitPython
PyPDF2
rank_bm25
requests
numpy